remove unused code, add hook for getting community stats
