--- app/src/marketplace/components/ProductTile/ActionsDropdown/actionsDropdown.pcss
@@ -1,26 +1,0 @@
-.root {
-  font-size: 1rem;
-}
-
-.toggle {
-  padding: 1em;
-  position: absolute;
-  right: 0;
-  top: 0;
-  z-index: 1;
-  cursor: pointer;
-}
-
-.menu {
-  border: 0;
-  border-radius: 0;
-  box-shadow: 0 0 10px rgba(0, 0, 0, 0.35);
-  font-size: 1em;
-  line-height: 2;
-  width: 160px;
-}
-
-.item {
-  color: #303030;
-  cursor: pointer;
-}
