+++ app/src/userpages/components/ProductsPage/index.jsx
@@ -25,12 +25,11 @@ import DocsShortcuts from '$userpages/components/DocsShortcuts'
 import ListContainer from '$shared/components/Container/List'
 import TileGrid from '$shared/components/TileGrid'
 import { isCommunityProduct } from '$mp/utils/product'
-import { getAllCommunityStats } from '$mp/modules/communityProduct/actions'
-import { selectCommunityProducts, selectFetchingCommunityStats } from '$mp/modules/communityProduct/selectors'
 import type { ProductId, Product } from '$mp/flowtype/product-types'
 import useFilterSort from '$userpages/hooks/useFilterSort'
 import useCopy from '$shared/hooks/useCopy'
 import useModal from '$shared/hooks/useModal'
+import useCommunityStats from '$mp/modules/communityProduct/hooks/useCommunityStats'
 import routes from '$routes'
 
 import CreateProductModal from '$mp/containers/CreateProductModal'
@@ -180,27 +179,15 @@ const ProductsPage = () => {
     const products = useSelector(selectMyProductList)
     const fetching = useSelector(selectFetching)
     const dispatch = useDispatch()
-    const communityStats = useSelector(selectCommunityProducts)
-    const fetchingCommunityStats = useSelector(selectFetchingCommunityStats)
-
-    const members = useMemo(() => (
-        (communityStats || {}).reduce((result, { id, memberCount }) => {
-            if (!memberCount) { return result }
-
-            return {
-                ...result,
-                [id.toLowerCase()]: memberCount.total,
-            }
-        }, {})
-    ), [communityStats])
+    const { load: loadCommunityStats, members, fetching: fetchingCommunityStats } = useCommunityStats()
 
     useEffect(() => {
         dispatch(getMyProducts(filter))
     }, [dispatch, filter])
 
     useEffect(() => {
-        dispatch(getAllCommunityStats())
-    }, [dispatch])
+        loadCommunityStats()
+    }, [loadCommunityStats])
 
     return (
         <Layout
