+++ app/src/userpages/components/PurchasesPage/index.jsx
@@ -12,7 +12,6 @@ import links from '../../../links'
 import { getFilters } from '../../utils/constants'
 import { getMyPurchases, updateFilter, applyFilter } from '$mp/modules/myPurchaseList/actions'
 import { selectMyPurchaseList, selectSubscriptions, selectFetchingMyPurchaseList } from '$mp/modules/myPurchaseList/selectors'
-import { selectCommunityProducts, selectFetchingCommunityStats } from '$mp/modules/communityProduct/selectors'
 import Tile from '$shared/components/Tile'
 import { isActive } from '$mp/utils/time'
 import Search from '../Header/Search'
@@ -23,7 +22,7 @@ import ListContainer from '$shared/components/Container/List'
 import TileGrid from '$shared/components/TileGrid'
 import { isCommunityProduct } from '$mp/utils/product'
 import useFilterSort from '$userpages/hooks/useFilterSort'
-import { getAllCommunityStats } from '$mp/modules/communityProduct/actions'
+import useCommunityStats from '$mp/modules/communityProduct/hooks/useCommunityStats'
 
 import type { ProductSubscription } from '$mp/flowtype/product-types'
 
@@ -52,8 +51,8 @@ const PurchasesPage = () => {
     const subscriptions = useSelector(selectSubscriptions)
     const fetching = useSelector(selectFetchingMyPurchaseList)
     const dispatch = useDispatch()
-    const communityStats = useSelector(selectCommunityProducts)
-    const fetchingCommunityStats = useSelector(selectFetchingCommunityStats)
+
+    const { load: loadCommunityStats, members, fetching: fetchingCommunityStats } = useCommunityStats()
 
     useEffect(() => {
         dispatch(updateFilter(filter))
@@ -64,19 +63,8 @@ const PurchasesPage = () => {
     }, [dispatch, filter])
 
     useEffect(() => {
-        dispatch(getAllCommunityStats())
-    }, [dispatch])
-
-    const members = useMemo(() => (
-        (communityStats || {}).reduce((result, { id, memberCount }) => {
-            if (!memberCount) { return result }
-
-            return {
-                ...result,
-                [id.toLowerCase()]: memberCount.total,
-            }
-        }, {})
-    ), [communityStats])
+        loadCommunityStats()
+    }, [loadCommunityStats])
 
     return (
         <Layout
